pyproject.toml
[project]
name = "prompt-go"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.9.0",
    "asyncio>=3.4.3",
    "coloredlogs>=15.0.1",
    "pynput>=1.7.6",
    "pyperclip>=1.8.2",
    "pytest-asyncio>=0.21.0",
    "pytest>=7.4.0",
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
    "watchdog>=3.0.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
